remove all portal routes, put map info in root path
