+++ src/Frontend/Pages/App.tsx
@@ -9,15 +9,12 @@ import { createGlobalStyle } from "styled-components";
 import { Theme } from "../Components/Theme";
 import { LandingPageBackground } from "../Renderers/LandingPageCanvas";
 import dfstyles from "../Styles/dfstyles";
-import { MapInfoView } from "../Views/Portal/MapInfoView";
 import { EntryPage } from "./EntryPage";
-import { EventsPage } from "./EventsPage";
 import { GifMaker } from "./GifMaker";
 import { ShareArtifact } from "./ShareArtifact";
 import { SharePlanet } from "./SharePlanet";
 import { TestArtifactImages } from "./TestArtifactImages";
 import { TxConfirmPopup } from "./TxConfirmPopup";
-import { ValhallaPage } from "./ValhallaPage";
 
 const isProd = process.env.NODE_ENV === "production";
 
@@ -29,16 +26,12 @@ function App() {
       <Theme color="dark" scale="medium">
         <Router>
           <Switch>
-            <Redirect exact from="/" to="/portal/home/" />
-            <Route path="/" exact component={MapInfoView} />
             <Route path="/planet/:locationId" component={SharePlanet} />
-            <Route path="/events" component={EventsPage} />
             <Route path="/artifact/:artifactId" component={ShareArtifact} />
             <Route
               path="/wallet/:contract/:addr/:actionId/:balance/:method"
               component={TxConfirmPopup}
             />
-            <Route path="/valhalla" component={ValhallaPage} />
             {!isProd && <Route path="/images" component={TestArtifactImages} />}
             {!isProd && <Route path="/gifs" component={GifMaker} />}
             {!isProd && <Route path="/bg" component={LandingPageBackground} />}
